Cleanup

+++ ControlUnit/sensors/light_intensity.c
@@ -1,4 +1,5 @@
 #include "light_intensity.h"
+#include "../ports/adc.h"
 #include <avr/io.h>
 
 const uint16_t MINIMUM_LIGHT_INTENSITY = 0;
@@ -8,9 +9,9 @@ const uint16_t MAXIMUM_LIGHT_INTENSITY = 800;
 uint8_t get_light_intensity(void)
 {
 	//Gets the light_level of the assigned analog pin
-	uint32_t light_level = adc_read(PINC1);
+	uint16_t light_level = adc_read(PINC1);
 	//Converts the light intensity to a %amount
-	uint32_t light_intensity = (((light_level - MINIMUM_LIGHT_INTENSITY) * 100) / (MAXIMUM_LIGHT_INTENSITY - MINIMUM_LIGHT_INTENSITY));
+	uint16_t light_intensity = (((light_level - MINIMUM_LIGHT_INTENSITY) * 100) / (MAXIMUM_LIGHT_INTENSITY - MINIMUM_LIGHT_INTENSITY));
 	//If the light_level is above the max amount it can handle set the %amount to 100%
 	if (light_level > MAXIMUM_LIGHT_INTENSITY)
 	{
